backend/requirements.txt
fastapi 
uvicorn
uvicorn[standard]
websockets
google-genai
quart
quart-cors
hypercorn
python-dotenv